% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coord_capped_cart}
\alias{coord_capped_cart}
\title{Cartesian coordinates with capped axes}
\usage{
coord_capped_cart(
  xlim = NULL,
  ylim = NULL,
  expand = TRUE,
  top = NULL,
  left = "both",
  bottom = "both",
  right = NULL,
  gap = 0.01
)
}
\arguments{
\item{xlim, ylim}{Limits for the x and y axes.}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}

\item{top, left, bottom, right}{Either a function returned from
\code{\link[lemon]{capped_horizontal}} or \code{\link[lemon]{brackets_horizontal}}.
If string, it is assumed to be shorthand for
\code{capped_horizontal(capped)} or similar for vertical.}

\item{gap}{Both ends are \emph{always} capped by this proportion.
Usually a value between 0 and 1.}
}
\value{
A coordinate system object that can be added to a ggplot object.
}
\description{
A wrapper around \code{\link[lemon:coord_capped]{lemon::coord_capped_cart()}} creating a Cartesian
coordinate system with capped left and bottom axis lines by default.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(hp, mpg)) +
  geom_point() +
  coord_capped_cart()
}
}
