% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eriplot.R
\name{eriplot}
\alias{eriplot}
\title{Create a new ggplot object}
\usage{
eriplot(data = NULL, mapping = ggplot2::aes(), ...)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
Initializes a new \code{ggplot} object, prepending the class
\code{"eri_ggplot"}. This provides a hook for setting new styles.
}
\examples{
\dontrun{
library(eriplots)

eriplot(mtcars, aes(hp, mpg)) +
  geom_point()
}
}
