% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus-signs.R
\name{format_minus_signs}
\alias{format_minus_signs}
\title{Repair ggplot2's default minus-sign formatting}
\usage{
format_minus_signs(plot, ...)
}
\arguments{
\item{plot}{A ggplot object}

\item{...}{Unused}
}
\description{
\pkg{ggplot2} applies R's \code{format()} for continuous scales, meaning negative
numbers are formatted with hyphens instead of proper minus signs. This method
identifies the continuous scales for which the user has not customizes the labels
and fixes them to have proper minus signs.
}
\examples{
\dontrun{
library(ggplot2)

# Apply directly
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
format_minus_signs(p)

# Apply compositionally
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  format_minus_signs()
}
}
