% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus-signs.R
\name{ggplot_add.format_minus_signs}
\alias{ggplot_add.format_minus_signs}
\title{Make it possible to add this to ggplot2}
\usage{
\method{ggplot_add}{format_minus_signs}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add \code{object} to}

\item{object_name}{The name of the object to add}
}
\description{
Make it possible to add this to ggplot2
}
\keyword{internal}
