% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eriplot.R
\name{ggplot_build.eri_ggplot}
\alias{ggplot_build.eri_ggplot}
\title{Build a customized ggplot object}
\usage{
\method{ggplot_build}{eri_ggplot}(plot)
}
\arguments{
\item{plot}{ggplot object}
}
\description{
Build a customized ggplot object
}
\keyword{internal}
