% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_default_profile}
\alias{set_default_profile}
\title{Set the default profile for new figures}
\usage{
set_default_profile(profile = c("none", "presentation", "document"))
}
\arguments{
\item{profile}{One of "none", "presentation", and "document".}
}
\value{
The new profile name, invisibly.
}
\description{
Set the default profile for new figures
}
