% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_eri}
\alias{theme_eri}
\title{Clean baseline \pkg{ggplot2} theme}
\usage{
theme_eri(
  profile = NULL,
  base_size = NULL,
  base_family = NULL,
  base_line_size = NULL,
  base_rect_size = NULL,
  aspect = NULL
)
}
\arguments{
\item{profile}{One of "none", "presentation", or "document".
Sets the text size and other parameters to support the
selected document type.}

\item{base_size}{Font size. All inherited font sizes are
defined relative to this size. The base size is 11 pt by
default, following tidyverse conventions for interactive
figures. Smaller sizes, like 10 pt, are be better suited for
print or presented figures.}

\item{base_family}{Font family. If not specified then
\link[ggplot2:ggplot2-package]{ggplot2::ggplot2-package} default fonts are used.}

\item{base_line_size}{Base line width. If not specified, this
is computed such that the base line width at 10 pt is 0.3 mm.}

\item{base_rect_size}{Base rectangle size. If not specified,
this is computed such that the size at 10 pt is 0.3 mm.
\code{\link[ggplot2:ggtheme]{theme_minimal()}}.}

\item{aspect}{Aspect ratio of the plotting area (height
divided by width). If \code{NULL} (default), the plotting area
will be scaled and shaped to fit the available space. If
specified, the plotting area will remain fixed to the
provided ratio, potentially leaving empty space along the
figure's left and right or top and bottom edges.}
}
\value{
A ggplot2 \link[ggplot2:theme]{theme}.
}
\description{
A simple, flexible \pkg{ggplot2} theme based on
\code{\link[ggplot2:ggtheme]{theme_minimal()}}. Provides clean,
simple defaults.
}
\details{
As in other ggplot2 themes, baseline theme settings can be specified
via function arguments. These value are propagated downwards through
the theme, and as far as possible downstream element sizes, faces,
etc. are defined relative to these base values.
}
\examples{
\dontrun{
library(ggplot2)
library(eriplots)

# Use the theme as-is
theme_set(theme_eri())

# Customize the theme, e.g., use a different typeface
theme_set(theme_eri(base_family = "serif"))
}
}
