% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eriplots-package.R
\docType{package}
\name{eriplots-package}
\alias{eriplots}
\alias{eriplots-package}
\title{eriplots: Simple, clean plotting tools}
\description{
The eriplots package provides a collection of tools for creating clean and consistent plots using ggplot2 and base R graphics. The package also provides a standardized theme and tools for saving figures that make it easier to follow good practices.
}
\author{
\strong{Maintainer}: Tom Shafer \email{tom.shafer@elderresearch.com}

Authors:
\itemize{
  \item Robert Robison \email{robert.robison@elderresearch.com}
}

Other contributors:
\itemize{
  \item Elder Research \email{support@elderresearch.com} [copyright holder]
}

}
\keyword{internal}
